fix: based on review
