+++ scripts/compile-bootloader.sh
@@ -4,7 +4,7 @@
 # to recompile the bootloader when changing the version of cairo-lang.
 
 SCRIPT_DIR=$( cd -- "$( dirname -- "${BASH_SOURCE[0]}" )" &> /dev/null && pwd )
-CAIRO_VERSION="0.13.1"
+CAIRO_VERSION="0.13.0"
 
 CAIRO_LANG_DIR=$(readlink -f "${SCRIPT_DIR}/../dependencies/cairo-lang")
 
@@ -27,6 +27,10 @@ OUTPUT_DIR="$(readlink -f "${SCRIPT_DIR}/../resources")"
 mkdir -p "${OUTPUT_DIR}"
 COMPILED_BOOTLOADER="${OUTPUT_DIR}/bootloader-${CAIRO_VERSION}.json"
 
-cairo-compile "${CAIRO_LANG_DIR}/src/starkware/cairo/bootloaders/bootloader/bootloader.cairo" --output "${COMPILED_BOOTLOADER}" --cairo_path "${CAIRO_LANG_DIR}"/src
+cairo-compile \
+  "${CAIRO_LANG_DIR}/src/starkware/cairo/bootloaders/bootloader/bootloader.cairo" \
+  --output "${COMPILED_BOOTLOADER}" \
+  --cairo_path "${CAIRO_LANG_DIR}"/src \
+  --proof_mode
 
 echo "Compiled the bootloader to '${COMPILED_BOOTLOADER}'."
