--- src/hints/hint_processor.rs
@@ -1,124 +1,0 @@
-use std::any::Any;
-use std::collections::HashMap;
-
-use cairo_vm::hint_processor::builtin_hint_processor::builtin_hint_processor_definition::HintProcessorData;
-use cairo_vm::hint_processor::builtin_hint_processor::memcpy_hint_utils::exit_scope;
-use cairo_vm::hint_processor::hint_processor_definition::HintProcessorLogic;
-use cairo_vm::types::exec_scope::ExecutionScopes;
-use cairo_vm::vm::errors::hint_errors::HintError;
-use cairo_vm::vm::runners::cairo_runner::ResourceTracker;
-use cairo_vm::vm::vm_core::VirtualMachine;
-use cairo_vm::Felt252;
-
-use crate::hints::bootloader_hints::{
-    assert_is_composite_packed_output, assert_program_address,
-    compute_and_configure_fact_topologies, enter_packed_output_scope,
-    guess_pre_image_of_subtasks_output_hash, import_packed_output_schemas, is_plain_packed_output,
-    load_bootloader_config, prepare_simple_bootloader_input,
-    prepare_simple_bootloader_output_segment, restore_bootloader_output, save_output_pointer,
-    save_packed_outputs, set_packed_output_to_subtasks,
-};
-use crate::hints::codes::*;
-use crate::hints::execute_task_hints::{
-    allocate_program_data_segment, append_fact_topologies, call_task, load_program_hint,
-    validate_hash, write_return_builtins_hint,
-};
-use crate::hints::inner_select_builtins::select_builtin;
-use crate::hints::select_builtins::select_builtins_enter_scope;
-use crate::hints::simple_bootloader_hints::{
-    divide_num_by_2, prepare_task_range_checks, set_ap_to_zero, set_current_task,
-    set_tasks_variable,
-};
-
-pub struct BootloaderHintProcessor;
-
-impl BootloaderHintProcessor {
-    pub fn new() -> Self {
-        Self {}
-    }
-}
-
-impl HintProcessorLogic for BootloaderHintProcessor {
-    fn execute_hint(
-        &mut self,
-        vm: &mut VirtualMachine,
-        exec_scopes: &mut ExecutionScopes,
-        hint_data: &Box<dyn Any>,
-        _constants: &HashMap<String, Felt252>,
-    ) -> Result<(), HintError> {
-        let hint_data = hint_data
-            .downcast_ref::<HintProcessorData>()
-            .ok_or(HintError::WrongHintData)?;
-
-        let ids_data = &hint_data.ids_data;
-        let ap_tracking = &hint_data.ap_tracking;
-
-        match hint_data.code.as_str() {
-            BOOTLOADER_RESTORE_BOOTLOADER_OUTPUT => restore_bootloader_output(vm, exec_scopes),
-            BOOTLOADER_PREPARE_SIMPLE_BOOTLOADER_INPUT => {
-                prepare_simple_bootloader_input(exec_scopes)
-            }
-            BOOTLOADER_LOAD_BOOTLOADER_CONFIG => {
-                load_bootloader_config(vm, exec_scopes, ids_data, ap_tracking)
-            }
-            BOOTLOADER_ENTER_PACKED_OUTPUT_SCOPE => {
-                enter_packed_output_scope(vm, exec_scopes, ids_data, ap_tracking)
-            }
-            BOOTLOADER_SAVE_OUTPUT_POINTER => {
-                save_output_pointer(vm, exec_scopes, ids_data, ap_tracking)
-            }
-            BOOTLOADER_SAVE_PACKED_OUTPUTS => save_packed_outputs(exec_scopes),
-            BOOTLOADER_GUESS_PRE_IMAGE_OF_SUBTASKS_OUTPUT_HASH => {
-                guess_pre_image_of_subtasks_output_hash(vm, exec_scopes, ids_data, ap_tracking)
-            }
-            BOOTLOADER_PREPARE_SIMPLE_BOOTLOADER_OUTPUT_SEGMENT => {
-                prepare_simple_bootloader_output_segment(vm, exec_scopes, ids_data, ap_tracking)
-            }
-            BOOTLOADER_COMPUTE_FACT_TOPOLOGIES => {
-                compute_and_configure_fact_topologies(vm, exec_scopes)
-            }
-            BOOTLOADER_SET_PACKED_OUTPUT_TO_SUBTASKS => set_packed_output_to_subtasks(exec_scopes),
-            BOOTLOADER_IMPORT_PACKED_OUTPUT_SCHEMAS => import_packed_output_schemas(),
-            BOOTLOADER_IS_PLAIN_PACKED_OUTPUT => is_plain_packed_output(vm, exec_scopes),
-            BOOTLOADER_ASSERT_IS_COMPOSITE_PACKED_OUTPUT => {
-                assert_is_composite_packed_output(exec_scopes)
-            }
-            SIMPLE_BOOTLOADER_PREPARE_TASK_RANGE_CHECKS => {
-                prepare_task_range_checks(vm, exec_scopes, ids_data, ap_tracking)
-            }
-            SIMPLE_BOOTLOADER_SET_TASKS_VARIABLE => set_tasks_variable(exec_scopes),
-            SIMPLE_BOOTLOADER_DIVIDE_NUM_BY_2 => divide_num_by_2(vm, ids_data, ap_tracking),
-            SIMPLE_BOOTLOADER_SET_CURRENT_TASK => {
-                set_current_task(vm, exec_scopes, ids_data, ap_tracking)
-            }
-            SIMPLE_BOOTLOADER_ZERO => set_ap_to_zero(vm),
-            EXECUTE_TASK_ALLOCATE_PROGRAM_DATA_SEGMENT => {
-                allocate_program_data_segment(vm, exec_scopes, ids_data, ap_tracking)
-            }
-            EXECUTE_TASK_LOAD_PROGRAM => load_program_hint(vm, exec_scopes, ids_data, ap_tracking),
-            EXECUTE_TASK_VALIDATE_HASH => validate_hash(vm, exec_scopes, ids_data, ap_tracking),
-            EXECUTE_TASK_ASSERT_PROGRAM_ADDRESS => {
-                assert_program_address(vm, exec_scopes, ids_data, ap_tracking)
-            }
-            EXECUTE_TASK_CALL_TASK => call_task(vm, exec_scopes, ids_data, ap_tracking),
-            EXECUTE_TASK_WRITE_RETURN_BUILTINS => {
-                write_return_builtins_hint(vm, exec_scopes, ids_data, ap_tracking)
-            }
-            EXECUTE_TASK_EXIT_SCOPE => exit_scope(exec_scopes),
-            EXECUTE_TASK_APPEND_FACT_TOPOLOGIES => {
-                append_fact_topologies(vm, exec_scopes, ids_data, ap_tracking)
-            }
-            SELECT_BUILTINS_ENTER_SCOPE => {
-                select_builtins_enter_scope(vm, exec_scopes, ids_data, ap_tracking)
-            }
-            INNER_SELECT_BUILTINS_SELECT_BUILTIN => {
-                select_builtin(vm, exec_scopes, ids_data, ap_tracking)
-            }
-            unknown_hint_code => Err(HintError::UnknownHint(
-                unknown_hint_code.to_string().into_boxed_str(),
-            )),
-        }
-    }
-}
-
-impl ResourceTracker for BootloaderHintProcessor {}
